HBP-95: Moved encrypting functions to ABCUploader
